requirements.txt
# =============================================================================
# FlavorSnap - Core Dependencies
# =============================================================================
# Python 3.8+ required
#
# Install with: pip install -r requirements.txt
# For GPU support, install PyTorch with CUDA from https://pytorch.org/get-started
# =============================================================================

# --- Machine Learning & Deep Learning ---
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
numpy>=1.24.0,<2.0.0

# --- Image Processing ---
Pillow>=10.0.0,<11.0.0
opencv-python>=4.8.0,<5.0.0

# --- Web Framework & API ---
Flask>=3.0.0,<4.0.0
Flask-Cors>=4.0.0,<5.0.0
Flask-Limiter>=3.5.0,<4.0.0
Werkzeug>=3.0.0,<4.0.0

# --- Data Visualization & Dashboard ---
panel>=1.3.0,<2.0.0
plotly>=5.18.0,<6.0.0
pandas>=2.1.0,<3.0.0
matplotlib>=3.8.0,<4.0.0
seaborn>=0.13.0,<1.0.0

# --- ML Metrics & Evaluation ---
scikit-learn>=1.3.0,<2.0.0
scipy>=1.11.0,<2.0.0

# --- Monitoring & Observability ---
prometheus-client>=0.19.0,<1.0.0
psutil>=5.9.0,<6.0.0

# --- API Documentation ---
PyYAML>=6.0.0,<7.0.0

# --- HTTP Client (for test scripts) ---
requests>=2.31.0,<3.0.0

# --- Export Functionality ---
reportlab>=4.0.0,<5.0.0
